requirements.txt
altair==4.1.0
anyio==3.4.0
apiclient==1.0.4
argon2-cffi==21.2.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
branca==0.4.2
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
defusedxml==0.7.1
Fiona==1.8.20
folium==0.12.1.post1
fonttools==4.28.3
geopandas==0.10.2
gitdb==4.0.9
GitPython==3.1.24
google==3.0.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
googletrans==4.0.0rc1
gspread==5.0.0
gspread-dataframe==3.2.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-resources==5.4.0
ipywidgets==7.6.5
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.2.1
jupyter-server==1.13.1
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
mistune==0.8.4
MultiApp==1.0
multitasking==0.0.10
munch==2.5.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
Pillow==8.4.0
prometheus-client==0.12.0
protobuf==3.19.1
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pympler==0.9
pyparsing==3.0.6
pyproj==3.3.0
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
retry==0.9.2
rfc3986==1.5.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
self==2020.12.3
Send2Trash==1.8.0
setuptools-scm==6.3.2
Shapely==1.8.0
six
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
streamlit==1.2.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==3.5.2
yfinance==0.1.67
zipp==3.6.0
